refactor: simplify Optimize instrumentation
