+++ optimize/backend/src/main/java/io/camunda/optimize/service/db/report/ReportEvaluationHandler.java
@@ -7,12 +7,13 @@
  */
 package io.camunda.optimize.service.db.report;
 
+import static io.camunda.optimize.ErrorType.TOO_MANY_BUCKETS;
 import static io.camunda.optimize.MetricEnum.REPORT_LATENCY_METRIC;
 import static io.camunda.optimize.dto.optimize.ReportConstants.ALL_VERSIONS;
 import static io.camunda.optimize.service.util.ExceptionUtil.isTooManyBucketsException;
 import static java.util.stream.Collectors.mapping;
 
-import io.camunda.optimize.ReportMetrics;
+import io.camunda.optimize.OptimizeMetrics;
 import io.camunda.optimize.dto.optimize.DefinitionType;
 import io.camunda.optimize.dto.optimize.RoleType;
 import io.camunda.optimize.dto.optimize.TenantDto;
@@ -88,7 +89,7 @@ public abstract class ReportEvaluationHandler {
 
   public AuthorizedReportEvaluationResult evaluateReport(
       final ReportEvaluationInfo evaluationInfo) {
-    return ReportMetrics.recordLatency(
+    return OptimizeMetrics.recordLatency(
         REPORT_LATENCY_METRIC,
         evaluationInfo::getReport,
         () -> {
@@ -258,6 +259,7 @@ public abstract class ReportEvaluationHandler {
       if (isTooManyBucketsException(e)) {
         final AuthorizedReportDefinitionResponseDto authorizedReportDefinitionDto =
             new AuthorizedReportDefinitionResponseDto(evaluationInfo.getReport(), currentUserRole);
+        OptimizeMetrics.recordError(TOO_MANY_BUCKETS.getValue());
         throw new TooManyBucketsException(authorizedReportDefinitionDto, e);
       } else {
         throw e;
